cloud/Issue_Identifier/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
google-genai
